requirements.txt
Django==3.1.7
djangorestframework==3.12.4
Pillow==8.2.0
environs~=9.3.2
django-cors-headers==3.7.0
